requirements.txt
Flask==3.1.3
gunicorn==21.2.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.30
google-auth==2.23.0
requests==2.31.0